rpcdaemon fuzz: cleanup
