+++ cmd/test/fuzzer_diagnosis.cpp
@@ -1,3 +1,19 @@
+/*
+   Copyright 2023 The Silkworm Authors
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
 #include <execinfo.h>
 
 #include <cstdlib>
@@ -6,7 +22,7 @@
 
 #include <CLI/CLI.hpp>
 
-#include "fuzzer_test.cpp"
+#include <silkworm/silkrpc/test/execution_api_database.hpp>
 
 void printStackTrace() {
     void* trace[16];
@@ -69,13 +85,13 @@ int main(int argc, char* argv[]) {
     silkworm::rpc::http::Reply reply;
 
     try {
-        static auto db = InitializeTestBase();
-        static auto request_handler = new RpcApiTestBase<RequestHandler_ForTest>(db);
+        static auto db = silkworm::rpc::test::InitializeTestBase();
+        static auto request_handler = new silkworm::rpc::test::RpcApiTestBase<silkworm::rpc::test::RequestHandler_ForTest>(db);
 
         auto request_json = nlohmann::json::parse(input_str);
         std::cout << "Request: " << request_json.dump(4) << std::endl;
 
-        request_handler->run<&RequestHandler_ForTest::handle_request>(input_str, reply);
+        request_handler->run<&silkworm::rpc::test::RequestHandler_ForTest::handle_request>(input_str, reply);
 
         auto db_path = db->get_path();
         db->close();
